backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Nguyen-Quy <96980131+Nguyen-Quy@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
anyio = "3.6.2"
bcrypt = "4.0.1"
cffi = "1.15.1"
click = "8.1.3"
cryptography = "40.0.2"
ecdsa = "0.18.0"
fastapi = "0.95.1"
h11 = "0.14.0"
idna = "3.4"
motor = "2.5.1"
passlib = "1.7.4"
pyasn1 = "0.4.8"
pycparser = "2.21"
pydantic = "1.10.7"
pyjwt = "2.6.0"
pymongo = "3.13.0"
python-jose = "3.3.0"
rsa = "4.9"
six = "1.16.0"
sniffio = "1.3.0"
starlette = "0.26.1"
typing-extensions = "4.5.0"
uvicorn = "0.21.1"
python-multipart = "^0.0.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
